Fixed some bugs

+++ NeuralNetwork/Util.py
@@ -8,27 +8,29 @@ import matplotlib.pyplot as plt
 from config import *
 
 
-def clear_cache():
-
-    def core(path):
-        if os.path.isfile(path):
-            os.remove(path)
-
-    if CLEAR_CACHE:
-        core("Data/cache.dat")
+def get_cache_path(path):
+    return "Data/data.cache" if path is None else path[:path.rfind(".")] + ".cache"
 
 
-def get_cache():
+def clear_cache(path, clear=CLEAR_CACHE):
+    path = get_cache_path(path)
+    if clear and os.path.isfile(path):
+        os.remove(path)
+
+
+def get_cache(path):
+    path = get_cache_path(path)
     try:
-        with open("Data/cache.dat", "rb") as file:
+        with open(path, "rb") as file:
             nn_data = pickle.load(file)
         return nn_data
     except FileNotFoundError:
         return None
 
 
-def do_cache(data):
-    with open("Data/cache.dat", "wb") as file:
+def do_cache(path, data):
+    path = get_cache_path(path)
+    with open(path, "wb") as file:
         pickle.dump(data, file)
 
 
@@ -37,12 +39,12 @@ def data_cleaning(line):
     return list(map(lambda c: c.strip(), line.split(",")))
 
 
-def get_data():
-
+def get_data(path=None):
+    path = "Data/data.txt" if path is None else path
     categories = None
 
     x = []
-    with open("Data/data.txt", "r") as file:
+    with open(path, "r") as file:
         flag = None
         for line in file:
             if SKIP_FIRST and flag is None:
@@ -54,7 +56,7 @@ def get_data():
             tmp_x = []
             if not DATA_CLEANED:
                 if categories is None:
-                    categories = [{ "flag": 1, _l: 0 } for _l in line]
+                    categories = [{"flag": 1, _l: 0} for _l in line]
                 for i, _l in enumerate(line):
                     if not WHETHER_NUMERICAL[i]:
                         if _l in categories[i]:
@@ -93,14 +95,13 @@ def get_data():
     return np.array(x), y
 
 
-def get_and_cache_data():
-
-    clear_cache()
-    _data = get_cache()
+def get_and_cache_data(path=None):
+    clear_cache(path)
+    _data = get_cache(path)
 
     if _data is None:
-        x, y = get_data()
-        do_cache((x, y))
+        x, y = get_data(path)
+        do_cache(path, (x, y))
     else:
         x, y = _data
 
@@ -127,20 +128,57 @@ def get_graphs_from_logs():
         plt.close()
 
 
-def gen_xor(size, scale):
+def init_size_and_path(size, path):
+    size = int(size)
+    path = "Data/data.txt" if path is None else path
+    return size, path
+
+
+def gen_xor(size, scale, path=None):
+    size, path = init_size_and_path(size, path)
+    with open(path, "w") as file:
+        quarter_size = int(size / 4)
+        seq1 = np.random.random(size=quarter_size) * scale
+        seq2 = np.random.random(size=quarter_size) * scale
+        seq = list(zip(seq1, seq2))
+        x0 = [(str(_s1), str(_s2)) for _s1, _s2 in seq]
+        x1 = [(str(-_s1), str(_s2)) for _s1, _s2 in seq]
+        x2 = [(str(-_s1), str(-_s2)) for _s1, _s2 in seq]
+        x3 = [(str(_s1), str(-_s2)) for _s1, _s2 in seq]
+        for i, x in enumerate((x0, x1, x2, x3)):
+            file.write("\n".join([",".join(_x) + ",{}".format(i % 2) for _x in x]) + "\n")
+
+
+def gen_random(size, scale, path=None):
+    size, path = init_size_and_path(size, path)
     quarter_size = int(size / 4)
-    with open("Data/data.txt", "w") as file:
-        seq = np.random.random(size=quarter_size) * scale
-        y0 = [(str(_s), str(_s)) for _s in seq]
-        y1 = [(str(-_s), str(_s)) for _s in seq]
-        y2 = [(str(-_s), str(-_s)) for _s in seq]
-        y3 = [(str(_s), str(-_s)) for _s in seq]
-        for i, y in enumerate((y0, y1, y2, y3)):
-            file.write("\n".join([",".join(_y) + ",{}".format(i % 2) for _y in y]) + "\n")
+    with open(path, "w") as file:
+        xs = (2 * np.random.rand(size, 2) - 1) * scale
+        xs = list(map(lambda v: (str(v[0]), str(v[1])), xs))
+        ans = np.random.randint(2, size=quarter_size * 4)
+        file.write("\n".join([",".join(x) + ",{}".format(y) for x, y in zip(xs, ans)]) + "\n")
+
+
+def gen_spin(size, n_classes=2, path=None):
+    size, path = init_size_and_path(size, path)
+    dimension = 2
+    xs = np.zeros((size * n_classes, dimension))
+    ys = np.zeros(size * n_classes, dtype='uint8')
+    for j in range(n_classes):
+        ix = range(size * j, size * (j + 1))
+        r = np.linspace(0.0, 1, size)
+        t = np.array(
+            np.linspace(j * (n_classes + 1), (j + 1) * (n_classes + 1), size) +
+            np.array(np.random.random(size=size)) * 0.2)
+        xs[ix] = np.c_[r * np.sin(t), r * np.cos(t)]
+        ys[ix] = j
+
+    with open(path, "w") as file:
+        xs = list(map(lambda v: (str(v[0]), str(v[1])), xs))
+        file.write("\n".join([",".join(x) + ",{}".format(y) for x, y in zip(xs, ys)]) + "\n")
 
 
 class ProgressBar:
-
     def __init__(self, min_value=None, max_value=None, width=30):
         self._min, self._max = min_value, max_value
         self._task_length = int(max_value - min_value) if (
@@ -224,5 +262,6 @@ class ProgressBar:
         self._started = True
         self._flush()
 
+
 if __name__ == '__main__':
-    get_graphs_from_logs()
+    gen_spin(100)
